+++ src/miniscreen/onboarding/connection/state.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from helpers import get_image_file_path
+from .images import ConnectionImages
 
 
 @dataclass(eq=True)
@@ -8,10 +8,6 @@ class ConnectionState:
     eth0_ip: str = ""
     ptusb0_ip: str = ""
     connected_device_ip: str = ""
-
-    ETHERNET_GIF_PATH = get_image_file_path("lan.gif")
-    USB_GIF_PATH = get_image_file_path("usb.gif")
-    CONNECT_GIF_PATH = get_image_file_path("first_time_connect.gif")
 
     def is_connected(self):
         return self.ethernet_is_connected() or self.usb_is_connected()
@@ -24,10 +20,10 @@ class ConnectionState:
 
     def path_to_image(self):
         if self.usb_is_connected():
-            return self.USB_GIF_PATH
+            return ConnectionImages.USB_GIF_PATH
         elif self.ethernet_is_connected():
-            return self.ETHERNET_GIF_PATH
-        return self.CONNECT_GIF_PATH
+            return ConnectionImages.ETHERNET_GIF_PATH
+        return ConnectionImages.CONNECT_GIF_PATH
 
     def ip_address(self):
         return self.ptusb0_ip if self.ptusb0_ip else self.eth0_ip
