+++ src/miniscreen/onboarding/main.py
@@ -3,10 +3,9 @@ from time import sleep
 
 from pitop import Pitop
 
-from device import Device
+from connection.monitor import ConnectionMonitor
 from helpers import (
     draw_text,
-    play_animated_image_file,
     MARGIN_X,
     FIRST_LINE_Y,
     SECOND_LINE_Y,
@@ -15,25 +14,23 @@ from helpers import (
 
 
 miniscreen = Pitop().miniscreen
-device = Device()
+connection = ConnectionMonitor()
 previous_state = None
 
+
 while True:
-    current_state = device.state
+    current_state = connection.state
     if current_state != previous_state:
         # update miniscreen with current state
         previous_state = current_state
-        play_animated_image_file(miniscreen, current_state.path_to_image())
+        miniscreen.play_animated_image_file(current_state.path_to_image(),
+                                            loop=current_state.is_connected() is False,
+                                            background=current_state.is_connected() is False)
         if current_state.is_connected():
             image = miniscreen.image.copy()
             canvas = ImageDraw.Draw(image)
-            draw_text(canvas, text=str(device.username), xy=(MARGIN_X, FIRST_LINE_Y),)
-            draw_text(canvas, text=str(device.password), xy=(MARGIN_X, SECOND_LINE_Y),)
+            draw_text(canvas, text=str(connection.username), xy=(MARGIN_X, FIRST_LINE_Y),)
+            draw_text(canvas, text=str(connection.password), xy=(MARGIN_X, SECOND_LINE_Y),)
             draw_text(canvas, text=str(current_state.ip_address()), xy=(MARGIN_X, THIRD_LINE_Y),)
             miniscreen.display_image(image)
-    elif current_state.is_connected():
-        # connection information is already displayed, sleep for a bit...
-        sleep(0.5)
-    else:
-        # not connected, play "connect" GIF again
-        play_animated_image_file(miniscreen, current_state.CONNECT_GIF_PATH)
+    sleep(0.5)
